requirements.txt
anyio==3.6.2
APScheduler==3.6.3
asgiref==3.6.0
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==2.0.12
dj-database-url==2.0.0
Django==4.2.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
psycopg2-binary==2.9.6
python-telegram-bot==13.12
pytz==2023.3
requests==2.28.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tbb @ git+https://github.com/natnaelab/tbb.git@2b259b21d58a33f73b6cd06eb8c8ff36a0ecfc90
tornado==6.3.2
typing_extensions==4.5.0
tzlocal==5.0.1
urllib3==1.26.15
